+++ R/expr-string.R
@@ -22,8 +22,8 @@ namespace_expr_str <- function(x) {
 #' string, if given, e.g.: `"%F %T%.3f"` => [`pl$Datetime("ms")`][pl__Datetime].
 #' If no fractional second component is found then the default is `"us"` (microsecond).
 # TODO: link to data type docs
-#' @param dtype The data type to convert into. Can be either `pl$Date`,
-#' `pl$Datetime`, or `pl$Time`.
+#' @param dtype The data type to convert into. Can be either [`pl$Date`][DataType],
+#'   [`pl$Datetime`][pl__Datetime], or [`pl$Time`][DataType].
 #' @param format Format to use for conversion. Refer to
 #' [the chrono crate documentation](https://docs.rs/chrono/latest/chrono/format/strftime/index.html)
 #' for the full specification. Example: `"%Y-%m-%d %H:%M:%S"`.
@@ -51,13 +51,6 @@ namespace_expr_str <- function(x) {
 #' df <- pl$DataFrame(x = c("2020-01-01 01:00Z", "2020-01-01 02:00Z"))
 #'
 #' df$select(pl$col("x")$str$strptime(pl$Datetime(), "%Y-%m-%d %H:%M%#z"))
-#'
-#' # Auto infer format
-#' df$select(pl$col("x")$str$strptime(pl$Datetime()))
-#'
-#' # Datetime with timezone is interpreted as UTC timezone
-#' df <- pl$DataFrame(x = c("2020-01-01T01:00:00+09:00"))
-#' df$select(pl$col("x")$str$strptime(pl$Datetime()))
 #'
 #' # Dealing with different formats.
 #' df <- pl$DataFrame(
@@ -194,7 +187,6 @@ expr_str_to_time <- function(format = NULL, ..., strict = TRUE, cache = TRUE) {
 #' df <- pl$DataFrame(x = c("2020-01-01 01:00Z", "2020-01-01 02:00Z"))
 #'
 #' df$select(pl$col("x")$str$to_datetime("%Y-%m-%d %H:%M%#z"))
-#' df$select(pl$col("x")$str$to_datetime(time_unit = "ms"))
 expr_str_to_datetime <- function(
   format = NULL,
   ...,
@@ -207,10 +199,14 @@ expr_str_to_datetime <- function(
 ) {
   wrap({
     check_dots_empty0(...)
-    if (!is_polars_expr(ambiguous)) {
-      ambiguous <- arg_match0(ambiguous, c("raise", "earliest", "latest", "null")) |>
+
+    ambiguous <- if (!is_polars_expr(ambiguous)) {
+      arg_match0(ambiguous, c("raise", "earliest", "latest", "null")) |>
         as_polars_expr(as_lit = TRUE)
+    } else {
+      ambiguous
     }
+
     self$`_rexpr`$str_to_datetime(
       format = format,
       time_unit = time_unit,
@@ -484,11 +480,12 @@ expr_str_zfill <- function(length) {
 
 #' Convert a String column into a Decimal column
 #'
-#' This method infers the needed parameters `precision` and `scale`.
+#' `r lifecycle::badge("experimental")`
 #'
 #' @inheritParams rlang::args_dots_empty
-#' @param inference_length Number of elements to parse to determine the
-#' `precision` and `scale`.
+#' @param scale Number of digits after the comma to use for the decimals.
+#' @param inference_length `r lifecycle::badge("deprecated")`
+#'   Ignored.
 #' @inherit as_polars_expr return
 #'
 #' @examples
@@ -498,11 +495,39 @@ expr_str_zfill <- function(length) {
 #'     "12.90", "143.09", "143.9"
 #'   )
 #' )
-#' df$with_columns(numbers_decimal = pl$col("numbers")$str$to_decimal())
-expr_str_to_decimal <- function(..., inference_length = 100) {
+#' df$with_columns(numbers_decimal = pl$col("numbers")$str$to_decimal(scale = 2))
+expr_str_to_decimal <- function(..., scale, inference_length = deprecated()) {
   wrap({
     check_dots_empty0(...)
-    self$`_rexpr`$str_to_decimal(inference_length)
+    if (is_present(inference_length)) {
+      deprecate_warn(
+        c(
+          `!` = sprintf(
+            "%s with %s is deprecated and has no effect on execution.",
+            format_code("<expr>$str$to_decimal()"),
+            format_arg("inference_length")
+          )
+        )
+      )
+    }
+
+    # Python Polars does not allow `scale` to be empty,
+    # but avoiding breaking change for the API, this is needed.
+    if (is_missing(scale)) {
+      deprecate_warn(
+        c(
+          `!` = sprintf(
+            "%s without %s is deprecated and set %s automatically.",
+            format_code("<expr>$str$to_decimal()"),
+            format_arg("scale"),
+            format_code("scale = 0L")
+          )
+        )
+      )
+      scale <- 0L
+    }
+
+    self$`_rexpr`$str_to_decimal(scale)
   })
 }
 
@@ -625,10 +650,9 @@ expr_str_starts_with <- function(prefix) {
 #' Parse string values as JSON.
 #'
 #' @inheritParams rlang::args_dots_empty
-#' @param dtype The dtype to cast the extracted value to. If `NULL`, the dtype
-#' will be inferred from the JSON value.
-#' @param infer_schema_length How many rows to parse to determine the schema.
-#' If `NULL`, all rows are used.
+#' @param dtype The dtype to cast the extracted value to.
+#' @param infer_schema_length `r lifecycle::badge("deprecated")`
+#'   Ignored.
 #' @details
 #' Throw errors if encounter invalid json strings.
 #'
@@ -638,18 +662,42 @@ expr_str_starts_with <- function(prefix) {
 #'   json_val = c('{"a":1, "b": true}', NA, '{"a":2, "b": false}')
 #' )
 #'
-#' df$select(
-#'   pl$col("json_val")$str$json_decode()
-#' )$unnest("json_val")
-#'
 #' dtype <- pl$Struct(a = pl$UInt8, b = pl$Boolean)
 #' df$select(
 #'   pl$col("json_val")$str$json_decode(dtype)
 #' )$unnest("json_val")
-expr_str_json_decode <- function(dtype = NULL, ..., infer_schema_length = 100) {
+expr_str_json_decode <- function(dtype, ..., infer_schema_length = deprecated()) {
   wrap({
     check_dots_empty0(...)
-    self$`_rexpr`$str_json_decode(dtype = dtype$`_dt`, infer_schema_length = infer_schema_length)
+
+    # Python Polars does not allow `dtype` to be empty,
+    # but avoiding breaking change for the API, this is needed.
+    if (is_missing(dtype)) {
+      deprecate_warn(
+        c(
+          `!` = sprintf(
+            "%s without %s is deprecated and set %s automatically.",
+            format_code("<expr>$str$json_decode()"),
+            format_arg("dtype"),
+            format_code("dtype = pl$Struct()")
+          )
+        )
+      )
+      dtype <- pl$Struct()
+    }
+    if (is_present(infer_schema_length)) {
+      deprecate_warn(
+        c(
+          `!` = sprintf(
+            "%s with %s is deprecated and has no effect on execution.",
+            format_code("<expr>$str$json_decode()"),
+            format_arg("infer_schema_length")
+          )
+        )
+      )
+    }
+
+    self$`_rexpr`$str_json_decode(dtype = dtype$`_dt`)
   })
 }