+++ NEWS.md
@@ -2,6 +2,25 @@
 
 ## polars (development version)
 
+This is an update that corresponds to Python Polars 1.33.0, which includes significant internal changes.
+
+### Deprecations
+
+Some of the methods or arguments of expr have been deprecated.
+They still work the same way on series.
+
+#### Entire expr method
+
+- `<expr>$shrink_dtype()` (#1507).
+
+#### Arguments of expr method
+
+- `<expr>$list$to_struct()`'s first argument `n_field_strategy` (#1507).
+- `<expr>$str$json_decode()`'s first argument `dtype` must be specified (#1507).
+- `<expr>$str$json_decode()`'s `infer_schema_length` (#1507).
+- `<expr>$str$to_decimal()`'s `inference_length` (#1507).
+- `<expr>$str$to_decimal()`'s new `scale` argument must be specified (#1507).
+
 ## polars 1.2.1
 
 This is an update that corresponds to Python Polars 1.32.3.